refactor: removed EditEventModal and replaced it with form component
